Take no action upon random mention

+++ slackify/handlers.py
@@ -26,35 +26,37 @@ def handle_app_mention(slack_client, message, channel_id):
                 thread_ts=message["event_ts"],
             )
             return
+        if "link" or "unlink" in text:
+            if "link" in text:
+                logger.info(
+                    "responding to %s for linking channel %s",
+                    message["user"],
+                    channel_id,
+                )
+                token = spotify.spotify_database.generate_token(
+                    spotify.spotify_database.get_db(), channel_id
+                )
+                link = "/".join([Config.BASE_URL, "spotify/link", channel_id, token])
+                response = "Follow this link to link this channel: {}".format(link)
 
-        if "link" in text:
-            logger.info(
-                "responding to %s for linking channel %s", message["user"], channel_id
+            if "unlink" in text:
+                logger.info("responding to unlink request")
+                response = "Follow this link to unlink this channel: {}".format(
+                    "[LINK TBD]"
+                )
+            slack_client.api_call(
+                "chat.postEphemeral",
+                channel=message["channel"],
+                text=response,
+                user=message["user"],
             )
-            token = spotify.spotify_database.generate_token(
-                spotify.spotify_database.get_db(), channel_id
+            slack_client.api_call(
+                "reactions.add",
+                channel=message["channel"],
+                name="notes",
+                timestamp=message["event_ts"],
             )
-            link = "/".join([Config.BASE_URL, "spotify/link", channel_id, token])
-            response = "Follow this link to link this channel: {}".format(link)
-
-        if "unlink" in text:
-            logger.info("responding to unlink request")
-            response = "Follow this link to unlink this channel: {}".format(
-                "[LINK TBD]"
-            )
-        slack_client.api_call(
-            "chat.postEphemeral",
-            channel=message["channel"],
-            text=response,
-            user=message["user"],
-        )
-        slack_client.api_call(
-            "reactions.add",
-            channel=message["channel"],
-            name="notes",
-            timestamp=message["event_ts"],
-        )
-        return
+            return
 
     else:
         logger.info("ignoring mention")
